Backend/requirements.txt
alembic==1.12.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
branca==0.8.1
bs4==0.0.1
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
config==0.5.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
display==1.0.0
distlib==0.3.7
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
dnspython==2.4.2
executing==2.1.0
fastapi==0.116.1
filelock==3.12.4
fiona==1.9.6
flack==1.3.7
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-MySQLdb==2.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
folium==0.20.0
fonttools==4.53.1
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.0
h11==0.14.0
idna==3.4
ipython==8.28.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
narwhals==2.0.1
networkx==3.4.2
numpy==1.26.4
osmnx==2.0.2
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==3.11.0
plotly==6.2.0
prompt_toolkit==3.0.48
protobuf==4.21.12
pure_eval==0.2.3
pyarrow==16.1.0
pycryptodome==3.19.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.5.0
PyMySQL==1.1.0
pyodbc==5.2.0
pyogrio==0.10.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
streamlit==1.35.0
sympy==1.13.2
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.0.6
uvicorn==0.35.0
virtualenv==20.24.5
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.0
xyzservices==2025.4.0